test: split parser downstream audits
