+++ src/parsers/scan_pipeline_test_utils.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 use std::sync::Arc;
 
 use crate::assembly;
-use crate::models::FileInfo;
+use crate::models::{DatasourceId, FileInfo, TopLevelDependency};
 use crate::progress::{ProgressMode, ScanProgress};
 use crate::scanner::{TextDetectionOptions, collect_paths, process_collected};
 
@@ -33,6 +33,42 @@ pub(crate) fn strip_root_paths(files: &mut [FileInfo], scan_root: &Path) {
             entry.path = stripped.to_string_lossy().to_string();
         }
     }
+}
+
+pub(crate) fn assert_dependency_present(
+    dependencies: &[TopLevelDependency],
+    purl: &str,
+    datafile_suffix: &str,
+) {
+    assert!(
+        dependencies.iter().any(|dep| {
+            dep.purl.as_deref() == Some(purl) && dep.datafile_path.ends_with(datafile_suffix)
+        }),
+        "expected dependency {purl} from {datafile_suffix}, found: {:?}",
+        dependencies
+            .iter()
+            .map(|dep| (dep.purl.clone(), dep.datafile_path.clone()))
+            .collect::<Vec<_>>()
+    );
+}
+
+pub(crate) fn assert_file_links_to_package(
+    files: &[FileInfo],
+    suffix: &str,
+    package_uid: &str,
+    datasource_id: DatasourceId,
+) {
+    let file = files
+        .iter()
+        .find(|file| file.path.ends_with(suffix))
+        .unwrap_or_else(|| panic!("{suffix} should be scanned"));
+
+    assert!(file.for_packages.iter().any(|uid| uid == package_uid));
+    assert!(
+        file.package_data
+            .iter()
+            .any(|pkg_data| { pkg_data.datasource_id == Some(datasource_id) })
+    );
 }
 
 fn strip_root_prefix(path: &Path, root: &Path) -> Option<std::path::PathBuf> {
