+++ testdata/chef/basic/metadata.json.expected.json
@@ -26,6 +26,7 @@
     "repository_homepage_url": "https://supermarket.chef.io/cookbooks/301/versions/0.1.0/",
     "repository_download_url": "https://supermarket.chef.io/cookbooks/301/versions/0.1.0/download",
     "api_data_url": "https://supermarket.chef.io/api/v1/cookbooks/301/versions/0.1.0",
-    "datasource_id": "chef_cookbook_metadata_rb"
+    "datasource_id": "chef_cookbook_metadata_json",
+    "purl": "pkg:chef/301@0.1.0"
   }
 ]
