+++ testdata/chef/dependencies/metadata.rb.expected.json
@@ -34,6 +34,7 @@
     "repository_homepage_url": "https://supermarket.chef.io/cookbooks/build-essential/versions/8.2.1/",
     "repository_download_url": "https://supermarket.chef.io/cookbooks/build-essential/versions/8.2.1/download",
     "api_data_url": "https://supermarket.chef.io/api/v1/cookbooks/build-essential/versions/8.2.1",
-    "datasource_id": "chef_cookbook_metadata_rb"
+    "datasource_id": "chef_cookbook_metadata_rb",
+    "purl": "pkg:chef/build-essential@8.2.1"
   }
 ]
